src/requirements.txt
altgraph==0.17
biopython==1.78
click==7.1.2
future==0.18.2
numpy==1.19.4
pefile==2019.4.18
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
PyQt5==5.15.1
pyqt5-plugins==5.15.1.2.0.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.1.3
python-dotenv==0.15.0
pywin32-ctypes==0.2.0
qt5-applications==5.15.1.2
qt5-tools==5.15.1.1.0.1
